+++ src/KristofferStrube.Blazor.FileSystemAccess/FileSystemHandle.cs
@@ -32,12 +32,12 @@ public class FileSystemHandle
         return await JSReference.InvokeAsync<bool>("isSameEntry", other.JSReference);
     }
 
-    public async Task<PermissionState> QueryPermission(FileSystemHandlePermissionDescriptor? description = null)
+    public async Task<PermissionState> QueryPermissionAsync(FileSystemHandlePermissionDescriptor? description = null)
     {
         return await JSReference.InvokeAsync<PermissionState>("queryPermission", description);
     }
 
-    public async Task<PermissionState> RequestPermission(FileSystemHandlePermissionDescriptor? description = null)
+    public async Task<PermissionState> RequestPermissionAsync(FileSystemHandlePermissionDescriptor? description = null)
     {
         return await JSReference.InvokeAsync<PermissionState>("requestPermission", description);
     }